+++ stoik-exposed-processor/src/main/kotlin/io/bkbn/stoik/exposed/processor/ExposedProcessor.kt
@@ -1,5 +1,7 @@
 package io.bkbn.stoik.exposed.processor
 
+import com.google.devtools.ksp.KspExperimental
+import com.google.devtools.ksp.getAnnotationsByType
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
@@ -8,20 +10,26 @@ import com.google.devtools.ksp.symbol.ClassKind
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSFile
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
 import com.google.devtools.ksp.symbol.KSValueArgument
 import com.google.devtools.ksp.symbol.KSVisitorVoid
 import com.google.devtools.ksp.validate
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
+import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.asTypeName
 import com.squareup.kotlinpoet.ksp.KotlinPoetKspPreview
 import com.squareup.kotlinpoet.ksp.addOriginatingKSFile
+import com.squareup.kotlinpoet.ksp.toTypeName
 import com.squareup.kotlinpoet.ksp.writeTo
+import io.bkbn.stoik.exposed.Column
 import io.bkbn.stoik.exposed.Table
+import io.bkbn.stoik.exposed.VarChar
 import io.bkbn.stoik.exposed.processor.util.StringHelpers.snakeToUpperCamelCase
 
 @OptIn(KotlinPoetKspPreview::class)
@@ -72,25 +80,54 @@ class ExposedProcessor(
 
       val tableName = nameArgument.value as String
       val tableObjectName = tableName.snakeToUpperCamelCase().plus("Table")
+      val containingFile = classDeclaration.containingFile ?: error("Could not identify originating file :(")
+      val properties: Sequence<KSPropertyDeclaration> = classDeclaration.getAllProperties()
+        .filter { it.validate() }
 
-      fileBuilder.addType(TypeSpec.objectBuilder(tableObjectName).apply {
-        addOriginatingKSFile(classDeclaration.containingFile ?: error("Could not identify originating file :("))
-        superclass(ClassName("org.jetbrains.exposed.dao.id", "UUIDTable"))
-        addSuperclassConstructorParameter("%S", tableName)
+      fileBuilder.createTableObject(tableName, tableObjectName, containingFile, properties)
+    }
+  }
 
-        val properties: Sequence<KSPropertyDeclaration> = classDeclaration.getAllProperties()
-          .filter { it.validate() }
-
-        properties.forEach { property ->
-          // todo get column name from annotation
-          val fieldName = property.simpleName.asString()
-          val columnType = ClassName("org.jetbrains.exposed.sql", "Column").parameterizedBy(String::class.asTypeName())
-          addProperty(PropertySpec.builder(fieldName, columnType).apply {
-            initializer("varchar(%S, %L)", fieldName, DEFAULT_VARCHAR_SIZE)
-          }.build())
-        }
-
+  @OptIn(KspExperimental::class)
+  private fun FileSpec.Builder.createTableObject(
+    tableName: String,
+    tableObjectName: String,
+    containingFile: KSFile,
+    properties: Sequence<KSPropertyDeclaration>
+  ) = addType(TypeSpec.objectBuilder(tableObjectName).apply {
+    addOriginatingKSFile(containingFile)
+    superclass(ClassName("org.jetbrains.exposed.dao.id", "UUIDTable"))
+    addSuperclassConstructorParameter("%S", tableName)
+    properties.forEach { property ->
+      val columnAnnotation: Column? = property.getAnnotationsByType(Column::class).firstOrNull()
+      val fieldName = property.simpleName.asString()
+      val columnName = columnAnnotation?.name ?: fieldName
+      val columnType = property.toColumnType()
+      addProperty(PropertySpec.builder(fieldName, columnType).apply {
+        setColumnInitializer(columnName, property)
       }.build())
     }
+  }.build())
+
+  private fun KSPropertyDeclaration.toColumnType(): TypeName {
+    val columnBase = ClassName("org.jetbrains.exposed.sql", "Column")
+    return columnBase.parameterizedBy(type.toTypeName())
+  }
+
+  private fun PropertySpec.Builder.setColumnInitializer(columnName: String, property: KSPropertyDeclaration) {
+    when (property.type.toTypeName()) {
+      String::class.asTypeName() -> initializer("varchar(%S, %L)", columnName, determineVarCharSize(property))
+      Int::class.asTypeName() -> initializer("integer(%S)", columnName)
+      Long::class.asTypeName() -> initializer("long(%S)", columnName)
+      Boolean::class.asTypeName() -> initializer("bool(%S)", columnName)
+      Float::class.asTypeName() -> initializer("float(%S)", columnName)
+      else -> TODO()
+    }
+  }
+
+  @OptIn(KspExperimental::class)
+  private fun determineVarCharSize(property: KSPropertyDeclaration): Int {
+    val varCharAnnotation = property.getAnnotationsByType(VarChar::class).firstOrNull()
+    return varCharAnnotation?.size ?: DEFAULT_VARCHAR_SIZE
   }
 }