+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 12/30/2021
+### Added
+- Exposed support for `Int`, `Long`, `Boolean` and `Float`
+- Added ability to override varchar length
+
 ## [0.1.1] - 12/29/2021
 ### Changed
 - Fixed error in test caused by not inheriting the classpath when processing annotations